Add more methods to the .Interfaces assembly - Moved ServiceCollectionExtensions - Moved ServiceProviderExtensions - Moved ITypeActivator (but it'll be gone soon)
