+++ src/Microsoft.Framework.DependencyInjection.Interfaces/ITypeActivator.cs
