+++ src/Microsoft.Framework.DependencyInjection.Interfaces/ServiceProviderExtensions.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using Microsoft.Framework.DependencyInjection.Interfaces;
 using Microsoft.Framework.Internal;
 
 namespace Microsoft.Framework.DependencyInjection
